server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
cryptography==43.0.1
python-multipart==0.0.9
boto3==1.35.24
httpx==0.27.2
apscheduler==3.10.4
reportlab==4.2.2
Pillow==10.4.0
aiofiles==24.1.0
slowapi==0.1.9
